Move mock contracts into mocks/

+++ contracts/mocks/MockOracle.sol
@@ -2,14 +2,15 @@ pragma solidity 0.5.10;
 
 import "../interfaces/OracleInterface.sol";
 
-contract MockOracle is OracleInterface{
+
+contract MockOracle is OracleInterface {
     uint256 public price;
 
     constructor() public {
-        price = 5 * (10 ** 15);
+        price = 5 * (10**15);
     }
 
-    function getPrice(address asset) external view returns (uint) {
+    function getPrice(address asset) external view returns (uint256) {
         if (asset == address(0)) {
             return (10**18);
         }
